# server/api/openapi.yaml
openapi: 3.1.0
info:
  title: Household Objects — HHO Ecosystem API
  version: "0.1.0-d1.6+g3"
  summary: >-
    The public API contract for the HHO Ecosystem server. Authored incrementally, one endpoint
    group per SOW deliverable, in the same commit as the handler it describes (constitution P-4,
    FR-081). Nothing in this document exists that the router does not currently mount; nothing the
    router mounts is missing from this document.
  license:
    name: AGPL-3.0-only
    identifier: AGPL-3.0-only
servers:
  - url: /api/v1
...
paths:
  /status:
    parameters:
      - $ref: "#/components/parameters/ClientVersion"
    get:
      operationId: getStatus
      tags: [status]
      summary: Liveness, build version and schema version.
      security: []
      responses:
        "200":
          description: The process is alive.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
        default:
          description: ''
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "426":
          $ref: "#/components/responses/UpgradeRequired"
  /openapi.yaml:
    parameters:
      - $ref: "#/components/parameters/ClientVersion"
    get:
      operationId: getOpenAPISpec
      tags: [meta]
      summary: This document, served from the running instance (FR-081).
      security: []
      responses:
        "200":
          description: The contract, as YAML.
          content:
            application/yaml:
              schema:
                type: string
        default:
          description: ''
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "426":
          $ref: "#/components/responses/UpgradeRequired"
  /auth/register:
    parameters:
      - $ref: "#/components/parameters/ClientVersion"
    post:
      operationId: registerFirstRun
      tags: [auth]
      summary: First-run self-service registration (FR-001).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Account and group created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterResponse"
        "400":
          description: ''
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "409":
          description: ''
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "429":
          description: ''
          headers:
            Retry-After:
              schema:
                type: integer
                minimum: 1
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "426":
          $ref: "#/components/responses/UpgradeRequired"
  /auth/login:
    parameters:
      - $ref: "#/components/parameters/ClientVersion"
    post:
      operationId: login
      tags: [auth]
      summary: Password login, session issuance (FR-002).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login succeeded. The session cookie is set via `Set-Cookie`.
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          description: The request body was not valid JSON. `detail` is populated.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          description: ''
          headers:
            WWW-Authenticate:
              schema:
                type: string
                enum: ["Bearer"]
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "429":
          description: ''
          headers:
            Retry-After:
              schema:
                type: integer
                minimum: 1
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "426":
          $ref: "#/components/responses/UpgradeRequired"
  /auth/sessions:
    parameters:
      - $ref: "#/components/parameters/ClientVersion"
    get:
      operationId: listSessions
      tags: [auth]
      summary: List the caller's own web sessions (FR-004).
      security:
        - SessionCookie: []
        - DeviceToken: []
      responses:
        "200":
          description: The caller's sessions, most recently created first.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "426":
          $ref: "#/components/responses/UpgradeRequired"
  /auth/sessions/{sessionID}:
    parameters:
      - $ref: "#/components/parameters/ClientVersion"
    delete:
      operationId: revokeSession
      tags: [auth]
      summary: Revoke one of the caller's own sessions (FR-004).
      security:
        - SessionCookie: []
        - DeviceToken: []
      parameters:
        - $ref: "#/components/parameters/SessionID"
      responses:
        "204":
          description: The session no longer authenticates (or was already revoked).
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: ''
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "426":
          $ref: "#/components/responses/UpgradeRequired"
  /auth/me:
    parameters:
      - $ref: "#/components/parameters/ClientVersion"
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: The caller's own identity (T154c).
      security:
        - SessionCookie: []
        - DeviceToken: []
      responses:
        "200":
          description: The caller's own identity.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "426":
          $ref: "#/components/responses/UpgradeRequired"
  /auth/logout:
    parameters:
      - $ref: "#/components/parameters/ClientVersion"
    post:
      operationId: logout
      tags: [auth]
      summary: End the calling session (G5, FR-004).
      security:
        - SessionCookie: []
      responses:
        "204":
          description: ''
          headers:
            Set-Cookie:
              description: Clears `hho_session`.
              schema:
                type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "426":
          $ref: "#/components/responses/UpgradeRequired"
  /auth/device-tokens:
    parameters:
      - $ref: "#/components/parameters/ClientVersion"
    get:
      operationId: listDeviceTokens
      tags: [auth]
      summary: List the caller's own registered devices (FR-004).
      security:
        - SessionCookie: []
        - DeviceToken: []
      responses:
        "200":
          description: The caller's device tokens, most recently created first.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceTokenListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "426":
          $ref: "#/components/responses/UpgradeRequired"
    post:
      operationId: issueDeviceToken
      tags: [auth]
      summary: Mint a device token for the caller (FR-003).
      security:
        - SessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceTokenIssueRequest"
      responses:
        "201":
          description: Device token minted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceTokenIssueResponse"
        "400":
          description: ''
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          description: ''
          headers:
            WWW-Authenticate:
              schema:
                type: string
                enum: ["Bearer"]
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "426":
          $ref: "#/components/responses/UpgradeRequired"
  /auth/device-tokens/{deviceTokenID}:
    parameters:
      - $ref: "#/components/parameters/ClientVersion"
    delete:
      operationId: revokeDeviceToken
      tags: [auth]
      summary: Revoke one of the caller's own device tokens (FR-004).
      security:
        - SessionCookie: []
        - DeviceToken: []
      parameters:
        - $ref: "#/components/parameters/DeviceTokenID"
      responses:
        "204":
          description: The device token no longer authenticates (or was already revoked).
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: ''
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "426":
          $ref: "#/components/responses/UpgradeRequired"
  /invites:
    parameters:
      - $ref: "#/components/parameters/ClientVersion"
    get:
      operationId: listInvites
      tags: [invites]
      summary: List every invite issued inside the caller's group (FR-006). Owner-only.
      security:
        - SessionCookie: []
        - DeviceToken: []
      responses:
        "200":
          description: Every invite issued inside the caller's group, most recently created first.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviteListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "426":
          $ref: "#/components/responses/UpgradeRequired"
    post:
      operationId: createInvite
      tags: [invites]
      summary: Issue a new invite token for the caller's group (FR-006). Owner-only.
      security:
        - SessionCookie: []
        - DeviceToken: []
      responses:
        "201":
          description: Invite created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviteCreateResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "426":
          $ref: "#/components/responses/UpgradeRequired"
  /invites/{inviteID}:
    parameters:
      - $ref: "#/components/parameters/ClientVersion"
    delete:
      operationId: revokeInvite
      tags: [invites]
      summary: Revoke one invite issued inside the caller's group (FR-006). Owner-only.
      security:
        - SessionCookie: []
        - DeviceToken: []
      parameters:
        - $ref: "#/components/parameters/InviteID"
      responses:
        "204":
          description: The invite is no longer redeemable (or already was).
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: ''
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "426":
          $ref: "#/components/responses/UpgradeRequired"
  /invites/redeem:
    parameters:
      - $ref: "#/components/parameters/ClientVersion"
    post:
      operationId: redeemInvite
      tags: [invites]
      summary: Redeem a bearer invite token into new group membership (FR-006).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteRedeemRequest"
      responses:
        "201":
          description: ''
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviteRedeemResponse"
        "400":
          description: ''
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "404":
          description: ''
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "409":
          description: ''
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "429":
          description: ''
          headers:
            Retry-After:
              schema:
                type: integer
                minimum: 1
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "426":
          $ref: "#/components/responses/UpgradeRequired"
  /items:
    parameters:
      - $ref: "#/components/parameters/ClientVersion"
    post:
      operationId: createItem
      tags: [items]
      summary: Create an item (FR-010).
      security:
        - SessionCookie: []
        - DeviceToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Item"
      responses:
        "201":
          description: Item created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "400":
          description: The request body was not valid JSON, or `name` was empty. `detail` is populated.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "426":
          $ref: "#/components/responses/UpgradeRequired"
    get:
      operationId: listItems
      tags: [items]
      summary: List and search the caller's group's live items (FR-020, FR-021, FR-033).
      security:
        - SessionCookie: []
        - DeviceToken: []
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: location_id
          in: query
          schema:
            type: string
        - name: descendants
          in: query
          schema:
            type: boolean
            default: false
          examples:
            subtree:
              value: true
        - name: label_id
          in: query
          explode: true
          schema:
            type: array
            items:
              type: string
        - name: custom_field
          in: query
          explode: true
          schema:
            type: array
            items:
              type: string
        - name: warranty_status
          in: query
          schema:
            type: string
            enum: [none, any, active, expired, lifetime]
        - name: created_from
          in: query
          schema:
            type: integer
            format: int64
        - name: created_to
          in: query
          schema:
            type: integer
            format: int64
          description: Only items created at or before this instant (INCLUSIVE). See `created_from`.
        - name: updated_from
          in: query
          schema:
            type: integer
            format: int64
        - name: updated_to
          in: query
          schema:
            type: integer
            format: int64
          description: Only items last updated at or before this instant (INCLUSIVE). See `updated_from`.
        - name: sort
          in: query
          schema:
            type: string
            pattern: '^-?(name|quantity|created_at|updated_at)$'
          examples:
            name_ascending:
              value: name
            newest_first:
              value: -created_at
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
          description: Page size, 1–200. An out-of-range or unparsable value is a 400 (A91).
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Rows to skip. A negative or unparsable value is a 400.
      responses:
        "200":
          description: The caller's group's live items, most recently updated first (or as ordered by `sort`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemListResponse"
        "400":
          description: ''
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "426":
          $ref: "#/components/responses/UpgradeRequired"
  /items/{itemID}:
    parameters:
      - $ref: "#/components/parameters/ClientVersion"
    get:
      operationId: getItem
      tags: [items]
      summary: Get one item (FR-020).
      security:
        - SessionCookie: []
        - DeviceToken: []
      parameters:
        - $ref: "#/components/parameters/ItemID"
      responses:
        "200":
          description: The item.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: ''
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "426":
          $ref: "#/components/responses/UpgradeRequired"
    put:
      operationId: updateItem
      tags: [items]
      summary: Replace an item's writable fields (FR-090's optimistic concurrency).
      security:
        - SessionCookie: []
        - DeviceToken: []
      parameters:
        - $ref: "#/components/parameters/ItemID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItemUpdateRequest"
      responses:
        "200":
          description: Item updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "400":
          description: ''
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: ''
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "409":
          description: ''
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "426":
          $ref: "#/components/responses/UpgradeRequired"
    delete:
      operationId: deleteItem
      tags: [items]
      summary: Soft-delete an item and cascade the tombstone to its satellite rows (FR-019, A107).
      security:
        - SessionCookie: []
        - DeviceToken: []
      parameters:
        - $ref: "#/components/parameters/ItemID"
      responses:
        "204":
          description: The item, and its satellite rows, are tombstoned.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: ''
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "426":
          $ref: "#/components/responses/UpgradeRequired"
...
  /groups/members:
    parameters:
      - $ref: "#/components/parameters/ClientVersion"
    get:
      operationId: listGroupMembers
      tags: [groups]
      summary: List every live member of the caller's group (FR-105, D1.17, A115). Member-readable.
      security:
        - SessionCookie: []
        - DeviceToken: []
      responses:
        "200":
          description: Every live member of the caller's own group, oldest-joined first.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupMembersResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "426":
          $ref: "#/components/responses/UpgradeRequired"
...
  /items/{itemID}/labels:
    parameters:
      - $ref: "#/components/parameters/ItemID"
      - $ref: "#/components/parameters/ClientVersion"
    get:
      operationId: listItemLabels
      tags: [items, labels]
      summary: List the labels attached to an item.
      security:
        - SessionCookie: []
        - DeviceToken: []
      responses:
        "200":
          description: The labels currently on this item. Empty when the item is untagged.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabelListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: No such item — unknown, tombstoned or another tenant's, indistinguishably. No `detail`.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "426":
          $ref: "#/components/responses/UpgradeRequired"
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: hho_session
    DeviceToken:
      type: http
      scheme: bearer
  parameters:
...
    ClientVersion:
      name: X-HHO-Client-Version
      in: header
      required: false
      schema:
        type: string
        pattern: "^[0-9]+\\.[0-9]+\\.[0-9]+$"
...
  responses:
    UpgradeRequired:
      description: ''
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/UpgradeRequiredProblem"
...
  schemas:
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
          description: Short human-readable summary, fixed per `type`.
        status:
          type: integer
          description: Repeats the HTTP status code.
        detail:
          type: string
        request_id:
          type: string
...
    StatusResponse:
      type: object
      required: [status, version, schema_version]
      properties:
        status:
          type: string
          enum: ["ok"]
        version:
          type: string
          description: The build version. "dev" for local builds.
        schema_version:
          type: integer
          format: int64
          description: The database migration version reached at startup.
...